реализован хук для проверки авторизации пользователя
