+++ packages/client/src/hooks/index.ts
@@ -1,1 +1,2 @@
 export * from './reduxToolkitHooks';
+export { useAuthChecker } from './useAuthChecker';
